+++ app/Http/Requests/Admin/PostRequest.php
@@ -14,10 +14,10 @@ class PostRequest extends Request
 	public function rules()
 	{
 		return [
-			'slug'    => ['required', 'unique:blg_translated_posts,id' . $this->get('post_id')],
+			'slug'    => ['unique:blg_translated_posts,id' . $this->get('post_id'), 'min:3'],
 			'title'   => ['required', 'min:5'],
 			'locale'  => ['required', 'in:' . implode(',', config()->get('app.locales'))],
-			'status'  => ['required', 'in:draft,published'],
+			'status'  => ['required', 'in:draft,published,hidden'],
 			'content' => ['required', 'min:10'],
 			'tags'    => ['array', 'min:1'],
 			'files'   => 'array',