+++ resources/views/layouts/blog.blade.php
@@ -26,11 +26,15 @@
 		<img style="position: absolute; top: 0; right: 0; border: 0;" src="{{ asset('/images/fork_me.png')}}" alt="Fork me on GitHub">
 	</a>
 	<script src="{{ asset('/assets/blog/app.js')}}"></script>
+	<script src="{{ asset('/assets/plugins/highlightjs/highlight.pack.min.js')}}"></script>
+	<script>hljs.initHighlightingOnLoad();</script>
 	@yield('custom_scripts')
+	@if(strlen(env('GOOGLE_ANALYTICS')) > 0)
 	<script>
 	(function(i,s,o,g,r,a,m){i['GoogleAnalyticsObject']=r;i[r]=i[r]||function(){(i[r].q=i[r].q||[]).push(arguments)},i[r].l=1*new Date();a=s.createElement(o),m=s.getElementsByTagName(o)[0];a.async=1;a.src=g;m.parentNode.insertBefore(a,m)})(window,document,'script','//www.google-analytics.com/analytics.js','ga');
-	ga('create', 'UA-61164802-1', 'auto');
+	ga('create', '{{ env('GOOGLE_ANALYTICS') }}', 'auto');
 	ga('send', 'pageview');
 	</script>
+	@endif
 </body>
 </html>