+++ changelog.md
@@ -1,5 +1,22 @@
 # Change Log
-All notable changes to this project will be documented in this file.
+All notable changes made to this project will be documented in this file.
+
+## [0.3.6] - [2016-01-07]
+### Added
+- Draft post preview for admin
+- Ability to generate slugs from title
+- Global exception handling
+- Helpers
+- Hidden post type
+
+### Change
+- Default env file
+- Some methods to helpers
+
+### Fixed
+- Redirect after file delete
+- Tag creation
+- Google Analytics id loaded from .env
 
 ## [0.3.5] - [2015-10-29]
 ### Added
@@ -58,7 +75,7 @@ All notable changes to this project will be documented in this file.
 - The ability to display drafts
 - Syntax highlighter in Safari
 
-## 0.2.0 - 2015-05-27
+## [0.2.0] - [2015-05-27]
 ### Added
 - Changelog
 - Permitted ip check on authorization
@@ -78,7 +95,7 @@ All notable changes to this project will be documented in this file.
 - jQuery in Front-End
 - Bootstrap JS in Front-End
 
-## 0.1.0 - 2015-05-19
+## [0.1.0] - [2015-05-19]
 ### Added
 - Authorization
 - "CRUD" actions to the post